Extract entries as mapped cards
